double-precision 개선

+++ Project2/material.h
@@ -1,22 +1,22 @@
 #pragma once
 #include "vec.h"
 
-typedef Vec3<float> Vec3f;
-typedef Vec4<float> Vec4f;
+typedef Vec3<double> Vec3d;
+typedef Vec4<double> Vec4d;
 
 class Material {
 private:
-	Vec4f color;			// RGBA (RGB + opacity) in [0,1]^4
-	float refraction_index;
-	float mirror;
+	Vec4d color;			// RGBA (RGB + opacity) in [0,1]^4
+	double refraction_index;
+	double mirror;
 
 public:
 	//constructor
-	Material() : color(Vec4f()), refraction_index(0), mirror(1) {};
-	Material(Vec4f _color) : refraction_index(0), mirror(1) { this->color = _color; };
-	Material(float _refration_index) : color(Vec4f()), mirror(1) { this->refraction_index = _refration_index; };
-	Material(Vec4f _color, float _refration_index) { this->color = _color; this->refraction_index = _refration_index; };
-	Material(Vec4f _color, float _refration_index, float _mirror) { this->color = _color; this->refraction_index = _refration_index; this->mirror = _mirror; };
+	Material() : color(Vec4d()), refraction_index(0), mirror(1) {};
+	Material(Vec4d _color) : refraction_index(0), mirror(1) { this->color = _color; };
+	Material(double _refration_index) : color(Vec4d()), mirror(1) { this->refraction_index = _refration_index; };
+	Material(Vec4d _color, double _refration_index) { this->color = _color; this->refraction_index = _refration_index; };
+	Material(Vec4d _color, double _refration_index, double _mirror) { this->color = _color; this->refraction_index = _refration_index; this->mirror = _mirror; };
 	Material(const Material& _material) {
 		this->color = _material.color;
 		this->refraction_index = _material.refraction_index;
@@ -32,13 +32,13 @@ public:
 	}
 
 	//getter and setter function
-	Vec3f getcolor() { return this->color; }
-	float getrefraction_index() { return this->refraction_index; }
-	float getopacity() { return this->color[3]; }
-	float getmirror() { return this->mirror; }
+	Vec3d getcolor() { return this->color; }
+	double getrefraction_index() { return this->refraction_index; }
+	double getopacity() { return this->color[3]; }
+	double getmirror() { return this->mirror; }
 
-	void setcolor(Vec3f _color) { this->color = _color; }
-	void setrefraction_index(float _refraction_index) { this->refraction_index = _refraction_index; }
-	void setopacity(float _opacity) { this->color[3] = _opacity; }
-	void setmirror(float _mirror) { this->mirror = _mirror; }
+	void setcolor(Vec3d _color) { this->color = _color; }
+	void setrefraction_index(double _refraction_index) { this->refraction_index = _refraction_index; }
+	void setopacity(double _opacity) { this->color[3] = _opacity; }
+	void setmirror(double _mirror) { this->mirror = _mirror; }
 };
